refactor(Attr): Avoid the need for `elements.js`
