+++ lib/jsdom/living/attributes.js
@@ -1,7 +1,5 @@
 "use strict";
 const DOMException = require("domexception/webidl2js-wrapper");
-const { appendAttribute, changeAttribute } = require("./elements.js");
-const attrGenerated = require("./generated/Attr");
 const { asciiLowercase } = require("./helpers/strings");
 const { HTML_NS } = require("./helpers/namespaces");
 const { queueAttributeMutationRecord } = require("./helpers/mutation-observers");
@@ -23,6 +21,42 @@ exports.hasAttributeByNameNS = function (element, namespace, localName) {
   return element._attributeList.some(attribute => {
     return attribute._localName === localName && attribute._namespace === namespace;
   });
+};
+
+// https://dom.spec.whatwg.org/#concept-element-attributes-change
+exports.changeAttribute = (element, attribute, value) => {
+  const { _localName, _namespace, _value } = attribute;
+  queueAttributeMutationRecord(element, _localName, _namespace, _value);
+
+  const oldValue = attribute._value;
+  attribute._value = value;
+
+  // Run jsdom hooks; roughly correspond to spec's "An attribute is set and an attribute is changed."
+  element._attrModified(attribute._qualifiedName, value, oldValue);
+};
+
+// https://dom.spec.whatwg.org/#concept-element-attributes-append
+exports.appendAttribute = function (element, attribute) {
+  const { _localName, _namespace } = attribute;
+  queueAttributeMutationRecord(element, _localName, _namespace, null);
+
+  const attributeList = element._attributeList;
+
+  attributeList.push(attribute);
+  attribute._element = element;
+
+  // Sync name cache
+  const name = attribute._qualifiedName;
+  const cache = element._attributesByNameMap;
+  let entry = cache.get(name);
+  if (!entry) {
+    entry = [];
+    cache.set(name, entry);
+  }
+  entry.push(attribute);
+
+  // Run jsdom hooks; roughly correspond to spec's "An attribute is set and an attribute is added."
+  element._attrModified(name, attribute._value, null);
 };
 
 exports.removeAttribute = function (element, attribute) {
@@ -159,7 +193,7 @@ exports.setAttribute = function (element, attr) {
   if (oldAttr !== null) {
     exports.replaceAttribute(element, oldAttr, attr);
   } else {
-    appendAttribute(element, attr);
+    exports.appendAttribute(element, attr);
   }
 
   return oldAttr;
@@ -177,18 +211,22 @@ exports.setAttributeValue = function (element, localName, value, prefix, namespa
 
   const attribute = exports.getAttributeByNameNS(element, namespace, localName);
   if (attribute === null) {
-    const newAttribute = attrGenerated.createImpl(element._globalObject, [], {
-      namespace,
-      namespacePrefix: prefix,
-      localName,
-      value
-    });
-    appendAttribute(element, newAttribute);
+    const newAttribute = element._ownerDocument._createAttribute(localName, namespace, prefix, value);
+    exports.appendAttribute(element, newAttribute);
 
     return;
   }
 
-  changeAttribute(element, attribute, value);
+  exports.changeAttribute(element, attribute, value);
+};
+
+exports.setAnExistingAttributeValue = (attribute, value) => {
+  const element = attribute._element;
+  if (element === null) {
+    attribute._value = value;
+  } else {
+    exports.changeAttribute(element, attribute, value);
+  }
 };
 
 exports.removeAttributeByName = function (element, name) {
