+++ lib/jsdom/living/attributes/Attr-impl.js
@@ -1,9 +1,7 @@
 "use strict";
 
-const {
-  setAnExistingAttributeValue,
-  implementation: NodeImpl
-} = require("../nodes/Node-impl.js");
+const { setAnExistingAttributeValue } = require("../attributes.js");
+const NodeImpl = require("../nodes/Node-impl.js").implementation;
 const { ATTRIBUTE_NODE } = require("../node-type.js");
 
 exports.implementation = class AttrImpl extends NodeImpl {
