+++ lib/jsdom/living/node.js
@@ -1,5 +1,5 @@
 "use strict";
-const { appendAttribute } = require("./elements.js");
+const { appendAttribute } = require("./attributes.js");
 const { cloningSteps, domSymbolTree } = require("./helpers/internal-constants");
 const NODE_TYPE = require("./node-type");
 const orderedSetParse = require("./helpers/ordered-set").parse;
@@ -7,7 +7,7 @@ const { asciiCaseInsensitiveMatch, asciiLowercase } = require("./helpers/strings
 const { HTML_NS, XMLNS_NS } = require("./helpers/namespaces");
 const HTMLCollection = require("./generated/HTMLCollection");
 
-module.exports.clone = function (node, document, cloneChildren) {
+exports.clone = (node, document, cloneChildren) => {
   if (document === undefined) {
     document = node._ownerDocument;
   }
@@ -31,7 +31,9 @@ module.exports.clone = function (node, document, cloneChildren) {
     case NODE_TYPE.ELEMENT_NODE:
       copy = document._createElementWithCorrectElementInterface(node._localName, node._namespaceURI);
       copy._prefix = node._prefix;
-      copyAttributeList(node, copy, document);
+      for (const attribute of node._attributeList) {
+        appendAttribute(copy, exports.clone(attribute, document));
+      }
       break;
 
     case NODE_TYPE.ATTRIBUTE_NODE:
@@ -65,7 +67,7 @@ module.exports.clone = function (node, document, cloneChildren) {
 
   if (cloneChildren) {
     for (const child of domSymbolTree.childrenIterator(node)) {
-      const childCopy = module.exports.clone(child, document, true);
+      const childCopy = exports.clone(child, document, true);
       copy._append(childCopy);
     }
   }
@@ -73,24 +75,9 @@ module.exports.clone = function (node, document, cloneChildren) {
   return copy;
 };
 
-function copyAttributeList(sourceElement, destElement, document) {
-  // Needed by https://dom.spec.whatwg.org/#concept-node-clone
-
-  for (const sourceAttr of sourceElement._attributeList) {
-    const destAttr = document._createAttribute(
-      sourceAttr._localName,
-      sourceAttr._namespace,
-      sourceAttr._namespacePrefix,
-      sourceAttr._value
-    );
-
-    appendAttribute(destElement, destAttr);
-  }
-}
-
 // For the following, memoization is not applied here since the memoized results are stored on `this`.
 
-module.exports.listOfElementsWithClassNames = (classNames, root) => {
+exports.listOfElementsWithClassNames = (classNames, root) => {
   // https://dom.spec.whatwg.org/#concept-getElementsByClassName
 
   const classes = orderedSetParse(classNames);
@@ -130,7 +117,7 @@ module.exports.listOfElementsWithClassNames = (classNames, root) => {
   });
 };
 
-module.exports.listOfElementsWithQualifiedName = (qualifiedName, root) => {
+exports.listOfElementsWithQualifiedName = (qualifiedName, root) => {
   // https://dom.spec.whatwg.org/#concept-getelementsbytagname
 
   if (qualifiedName === "*") {
@@ -177,7 +164,7 @@ module.exports.listOfElementsWithQualifiedName = (qualifiedName, root) => {
   });
 };
 
-module.exports.listOfElementsWithNamespaceAndLocalName = (namespace, localName, root) => {
+exports.listOfElementsWithNamespaceAndLocalName = (namespace, localName, root) => {
   // https://dom.spec.whatwg.org/#concept-getelementsbytagnamens
 
   if (namespace === "") {
