+++ lib/jsdom/living/nodes/Element-impl.js
@@ -6,7 +6,6 @@ const idlUtils = require("../generated/utils");
 const NodeImpl = require("./Node-impl").implementation;
 const ParentNodeImpl = require("./ParentNode-impl").implementation;
 const ChildNodeImpl = require("./ChildNode-impl").implementation;
-const { appendAttribute, changeAttribute } = require("../elements.js");
 const attributes = require("../attributes");
 const namedPropertiesWindow = require("../named-properties-window");
 const NODE_TYPE = require("../node-type");
@@ -242,11 +241,11 @@ class ElementImpl extends NodeImpl {
 
     if (attribute === null) {
       const newAttr = attrGenerated.createImpl(this._globalObject, [], { localName: name, value });
-      appendAttribute(this, newAttr);
+      attributes.appendAttribute(this, newAttr);
       return;
     }
 
-    changeAttribute(this, attribute, value);
+    attributes.changeAttribute(this, attribute, value);
   }
 
   setAttributeNS(namespace, name, value) {
@@ -279,7 +278,7 @@ class ElementImpl extends NodeImpl {
     if (attribute === null) {
       if (force === undefined || force === true) {
         const newAttr = attrGenerated.createImpl(this._globalObject, [], { localName: qualifiedName, value: "" });
-        appendAttribute(this, newAttr);
+        attributes.appendAttribute(this, newAttr);
         return true;
       }
       return false;
