+++ lib/jsdom/living/nodes/Node-impl.js
@@ -7,7 +7,7 @@ const { simultaneousIterators } = require("../../utils");
 const NODE_TYPE = require("../node-type");
 const NODE_DOCUMENT_POSITION = require("../node-document-position");
 const { clone, locateNamespacePrefix, locateNamespace } = require("../node");
-const { changeAttribute } = require("../elements.js");
+const { setAnExistingAttributeValue } = require("../attributes.js");
 
 const NodeList = require("../generated/NodeList");
 
@@ -24,7 +24,6 @@ function isObsoleteNodeType(node) {
   return node.nodeType === NODE_TYPE.ENTITY_NODE ||
     node.nodeType === NODE_TYPE.ENTITY_REFERENCE_NODE ||
     node.nodeType === NODE_TYPE.NOTATION_NODE ||
-  //  node.nodeType === NODE_TYPE.ATTRIBUTE_NODE ||  // this is missing how do we handle?
     node.nodeType === NODE_TYPE.CDATA_SECTION_NODE;
 }
 
@@ -520,7 +519,7 @@ exports.implementation = class NodeImpl extends EventTargetImpl {
 
     switch (this.nodeType) {
       case NODE_TYPE.ATTRIBUTE_NODE: {
-        exports.setAnExistingAttributeValue(this, value);
+        setAnExistingAttributeValue(this, value);
         break;
       }
       case NODE_TYPE.TEXT_NODE:
@@ -583,7 +582,7 @@ exports.implementation = class NodeImpl extends EventTargetImpl {
       }
 
       case NODE_TYPE.ATTRIBUTE_NODE: {
-        exports.setAnExistingAttributeValue(this, value);
+        setAnExistingAttributeValue(this, value);
         break;
       }
 
@@ -1132,14 +1131,5 @@ exports.implementation = class NodeImpl extends EventTargetImpl {
     if (nodeImpl.nodeType === NODE_TYPE.TEXT_NODE) {
       this._childTextContentChangeSteps();
     }
-  }
-};
-
-exports.setAnExistingAttributeValue = (attribute, value) => {
-  const element = attribute._element;
-  if (element === null) {
-    attribute._value = value;
-  } else {
-    changeAttribute(element, attribute, value);
   }
 };
